+++ sechub-wrapper-prepare/src/main/java/com/mercedesbenz/sechub/wrapper/prepare/modules/skopeo/SkopeoInputValidator.java
@@ -1,9 +1,10 @@
-package com.mercedesbenz.sechub.wrapper.prepare.modules;
+package com.mercedesbenz.sechub.wrapper.prepare.modules.skopeo;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Pattern;
 
+import com.mercedesbenz.sechub.wrapper.prepare.modules.InputValidator;
 import org.springframework.stereotype.Component;
 
 @Component
