+++ sechub-wrapper-prepare/src/main/java/com/mercedesbenz/sechub/wrapper/prepare/prepare/PrepareWrapperPreparationService.java
@@ -45,6 +45,7 @@ public class PrepareWrapperPreparationService {
             return createAdapterExecutionResult(PrepareStatus.OK, SecHubMessageType.WARNING, "No Remote Configuration found.");
         }
 
+        // TODO: 16.05.24 laura for each schleife hier und in modulen raus? nur eins möglich?
         for (PrepareWrapperModule module : modules) {
             if (!module.isAbleToPrepare(context)) {
                 continue;
